requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.15.30
awscrt==0.19.19
Babel==2.9.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.4.0
chevron==0.13.1
click==8.1.7
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distlib==0.3.9
distro==1.9.0
docutils==0.16
ec2-hibinit-agent==1.0.8
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
gpg==1.15.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==2.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
netifaces==0.10.6
oauthlib==3.0.2
openai==1.52.2
packaging==24.1
platformdirs==4.3.6
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
pycparser==2.20
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
python-dotenv==1.0.1
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.32.3
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
sniffio==1.3.1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.36
support-info==1.0
systemd-python==235
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.25.10
virtualenv==20.27.0
wcwidth==0.2.5
Werkzeug==3.0.6
zipp==3.20.2
